Move playwright stuff to BrowserTestBase
