+++ src/Shared/BrowserTesting/src/Microsoft.AspNetCore.BrowserTesting.csproj
@@ -11,6 +11,7 @@
   <ItemGroup>
     <Reference Include="PlaywrightSharp" Condition="'$(TargetOsName)' != 'linux-musl'" />
     <Reference Include="PlaywrightSharp" ExcludeAssets="build" Condition="'$(TargetOsName)' == 'linux-musl'" />
+    <Reference Include="Microsoft.AspNetCore.Testing" />
   </ItemGroup>
 
 </Project>
